+++ src/viur/core/config.py
@@ -97,7 +97,7 @@ conf = Conf({
     # Database engine module
     "viur.db.engine": "viur.datastore",
 
-    # If enabled, trace any routing and decorations for debugging and insight
+    # If enabled, trace any routing, HTTPExceptions and decorations for debugging and insight
     "viur.debug.trace": False,
     # If enabled, user-generated exceptions from the viur.core.errors module won't be caught and handled
     "viur.debug.traceExceptions": False,