+++ src/viur/core/tasks.py
@@ -218,15 +218,16 @@ class TaskHandler(Module):
             if not conf["viur.instance.is_dev_server"] or os.getenv("TASKS_EMULATOR") is None:
                 logging.critical('Detected an attempted XSRF attack. This request did not originate from Task Queue.')
             raise errors.Forbidden()
+
         # Check if the retry count exceeds our warning threshold
         retryCount = req.headers.get("X-Appengine-Taskretrycount", None)
-        if retryCount:
-            if int(retryCount) == self.retryCountWarningThreshold:
-                from viur.core import email
-                email.sendEMailToAdmins("Deferred task retry count exceeded warning threshold",
-                                        "Task %s will now be retried for the %sth time." % (
-                                            req.headers.get("X-Appengine-Taskname", ""),
-                                            retryCount))
+        if retryCount and int(retryCount) == self.retryCountWarningThreshold:
+            from viur.core import email
+            email.sendEMailToAdmins(
+                "Deferred task retry counter exceeded warning threshold",
+                f"""Task {req.headers.get("X-Appengine-Taskname", "")} is retried for the {retryCount}th time."""
+            )
+
         cmd, data = json.loads(req.body, object_hook=jsonDecodeObjectHook)
         funcPath, args, kwargs, env = data
         logging.debug(f"Call task {funcPath} with {cmd=} {args=} {kwargs=} {env=}")